+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [2.0.2] - Unreleased
 - [Fixed] Trip detail map disappearance on iOS while zooming or panning the map.
 - [Fixed] Errors caused by invalid trip data.
+- [Fixed] First location was not recorded in the trip.
 
 ## [2.0.1] - 2017-07-06
 - [Fixed] Set geolocation permission descriptions for iOS.